Add items dialog to systemsItemsTable
